requirements.txt
python-dotenv
openai
faiss-cpu
numpy
tiktoken